numba/numpy2.0 fixes

+++ openquake/hazardlib/geo/surface/multi.py
@@ -331,7 +331,7 @@ class MultiSurface(BaseSurface):
             A :class:`numpy.ndarray` instance with the Rx distance. Note that
             the Rx distance is directly taken from the GC2 t-coordinate.
         """
-        tut, _uut = self.tor.get_tu(mesh.lons, mesh.lats)
+        tut, _uut = self.tor.get_tu(mesh.lons.astype(float), mesh.lats.astype(float))
         rx = tut[0] if len(tut[0].shape) > 1 else tut
         return rx
 
@@ -342,7 +342,7 @@ class MultiSurface(BaseSurface):
             coordinates of the sites.
         """
         u_max = self.tor.get_u_max()
-        _tut, uut = self.tor.get_tu(mesh.lons, mesh.lats)
+        _tut, uut = self.tor.get_tu(mesh.lons.astype(float), mesh.lats.astype(float))
         ry0 = np.zeros_like(uut)
         ry0[uut < 0] = np.abs(uut[uut < 0])
         condition = uut > u_max
